# Minimize allocations when unpacking `TimeZone`s from cache (updated)