--- a/src/types/timezone.jl
+++ b/src/types/timezone.jl
@@ -1,37 +1,29 @@
-# Retains the compiled tzdata in memory. Read-only access to the cache is thread-safe and
-# any changes to this structure can result in inconsistent behaviour. Do not access this
-# object directly, instead use `_get_tz_cache_entry()` to access the cache content.
-#
 # Use a separate cache for FixedTimeZone (which is `isbits`) so the container is concretely
 # typed and we avoid allocating a FixedTimeZone every time we get one from the cache.
-const _FTZ_CACHE = Dict{String,Tuple{FixedTimeZone,Class}}()
-const _VTZ_CACHE = Dict{String,Tuple{VariableTimeZone,Class}}()
-const _TZ_CACHE_LOCK = ReentrantLock()
-const _TZ_CACHE_INITIALIZED = Threads.Atomic{Bool}(false)
-
-function _init_tz_cache()
-    # Write out our compiled tzdata representations into a scratchspace
-    desired_version = TZData.tzdata_version()
-
-    _COMPILED_DIR[] = if desired_version == TZJData.TZDATA_VERSION
-        TZJData.ARTIFACT_DIR
-    else
-        TZData.build(desired_version, _scratch_dir())
-    end
-
-    # Load the pre-computed TZData into memory.
-    return _reload_tz_cache(_COMPILED_DIR[])
+struct TimeZoneCache
+    ftz::Dict{String,Tuple{FixedTimeZone,Class}}
+    vtz::Dict{String,Tuple{VariableTimeZone,Class}}
+    lock::ReentrantLock
+    initialized::Threads.Atomic{Bool}
 end
 
-function _reload_tz_cache(compiled_dir::AbstractString)
-    _reload_tz_cache!(_FTZ_CACHE, _VTZ_CACHE, compiled_dir)
-    !isempty(_FTZ_CACHE) && !isempty(_VTZ_CACHE) || error("Cache remains empty after loading")
-    return nothing
+TimeZoneCache() = TimeZoneCache(Dict(), Dict(), ReentrantLock(), Threads.Atomic{Bool}(false))
+
+# Retains the compiled tzdata in memory. Read-only access to the cache is thread-safe and
+# any changes to this structure can result in inconsistent behaviour. Do not access this
+# object directly, instead use `get` to access the cache content.
+const _TZ_CACHE = TimeZoneCache()
+
+function Base.copy!(dst::TimeZoneCache, src::TimeZoneCache)
+    copy!(dst.ftz, src.ftz)
+    copy!(dst.vtz, src.vtz)
+    dst.initialized[] = src.initialized[]
+    return dst
 end
 
-function _reload_tz_cache!(ftz_cache::AbstractDict, vtz_cache::AbstractDict, compiled_dir::AbstractString)
-    empty!(ftz_cache)
-    empty!(vtz_cache)
+function reload!(cache::TimeZoneCache, compiled_dir::AbstractString=_COMPILED_DIR[])
+    empty!(cache.ftz)
+    empty!(cache.vtz)
     check = Tuple{String,String}[(compiled_dir, "")]
 
     for (dir, partial) in check
@@ -47,34 +39,53 @@ function _reload_tz_cache!(ftz_cache::AbstractDict, vtz_cache::AbstractDict, com
                 tz, class = open(TZJFile.read, path, "r")(name)
 
                 if tz isa FixedTimeZone
-                    ftz_cache[name] = (tz, class)
+                    cache.ftz[name] = (tz, class)
                 elseif tz isa VariableTimeZone
-                    vtz_cache[name] = (tz, class)
+                    cache.vtz[name] = (tz, class)
                 else
                     error("Unhandled TimeZone class encountered: $(typeof(tz))")
                 end
             end
         end
     end
-    return nothing
+
+    !isempty(cache.ftz) && !isempty(cache.vtz) || error("Cache remains empty after loading")
+
+    return cache
 end
 
-function _get_tz_cache_entry(body::Function, name::AbstractString)
-    if !_TZ_CACHE_INITIALIZED[]
-        lock(_TZ_CACHE_LOCK) do
-            if !_TZ_CACHE_INITIALIZED[]
-                _init_tz_cache()
-                _TZ_CACHE_INITIALIZED[] = true
+function Base.get(body::Function, cache::TimeZoneCache, name::AbstractString)
+    if !cache.initialized[]
+        lock(cache.lock) do
+            if !cache.initialized[]
+                _initialize()
+                reload!(cache)
+                cache.initialized[] = true
             end
         end
     end
 
-    return get(_FTZ_CACHE, name) do
-        get(_VTZ_CACHE, name) do
+    return get(cache.ftz, name) do
+        get(cache.vtz, name) do
             body()
         end
     end
 end
+
+function _initialize()
+    # Write out our compiled tzdata representations into a scratchspace
+    desired_version = TZData.tzdata_version()
+
+    _COMPILED_DIR[] = if desired_version == TZJData.TZDATA_VERSION
+        TZJData.ARTIFACT_DIR
+    else
+        TZData.build(desired_version, _scratch_dir())
+    end
+
+    return nothing
+end
+
+_reload_tz_cache(compiled_dir::AbstractString) = reload!(_TZ_CACHE, compiled_dir)
 
 """
     TimeZone(str::AbstractString) -> TimeZone
@@ -117,7 +128,7 @@ US/Pacific (UTC-8/UTC-7)
 TimeZone(::AbstractString, ::Class)
 
 function TimeZone(str::AbstractString, mask::Class=Class(:DEFAULT))
-    tz, class = _get_tz_cache_entry(str) do
+    tz, class = get(_TZ_CACHE, str) do
         if occursin(FIXED_TIME_ZONE_REGEX, str)
             FixedTimeZone(str), Class(:FIXED)
         else
@@ -162,6 +173,6 @@ function istimezone(str::AbstractString, mask::Class=Class(:DEFAULT))
     end
 
     # Checks against pre-compiled time zones
-    class = _get_tz_cache_entry(() -> (UTC_ZERO, Class(:NONE)), str)[2]
+    class = get(() -> (UTC_ZERO, Class(:NONE)), _TZ_CACHE, str)[2]
     return mask & class != Class(:NONE)
 end