--- a/test/helpers.jl
+++ b/test/helpers.jl
@@ -47,43 +47,37 @@ show_compact = (io, args...) -> show(IOContext(io, :compact => true), args...)
 # Modified the internal TimeZones cache. Should only be used as part of testing and only is
 # needed when the data between the test tzdata version and the built tzdata versions differ.
 
-function add!(cache::Dict, t::Tuple{TimeZone,TimeZones.Class})
+function add!(dict::Dict, t::Tuple{TimeZone,TimeZones.Class})
     tz, class = t
     name = TimeZones.name(tz)
-    push!(cache, name => t)
+    push!(dict, name => t)
     return tz
 end
 
-function add!(cache::Dict, tz::VariableTimeZone)
+function add!(cache::TimeZones.TimeZoneCache, t::Tuple{T,TimeZones.Class}) where {T<:TimeZone}
+    dict = T == FixedTimeZone ? cache.ftz : cache.vtz
+    return add!(dict, t)
+end
+
+function add!(cache::TimeZones.TimeZoneCache, tz::VariableTimeZone)
     # Not all `VariableTimeZone`s are the STANDARD class. However, for testing purposes
     # the class doesn't need to be precise.
     class = TimeZones.Class(:STANDARD)
-    return add!(cache, (tz, class))
+    return add!(cache.vtz, (tz, class))
 end
 
-function add!(cache::Dict, tz::FixedTimeZone)
+function add!(cache::TimeZones.TimeZoneCache, tz::FixedTimeZone)
     class = TimeZones.Class(:FIXED)
-    return add!(cache, (tz, class))
+    return add!(cache.ftz, (tz, class))
 end
 
-function with_tz_cache(f, cache::Dict{String,Tuple{TimeZone,TimeZones.Class}})
-    old_ftz_cache = deepcopy(TimeZones._FTZ_CACHE)
-    old_vtz_cache = deepcopy(TimeZones._VTZ_CACHE)
-
-    # Split the contents of `cache` between the fixed and variable caches
-    # as appropriate.
-    empty!(TimeZones._FTZ_CACHE)
-    empty!(TimeZones._VTZ_CACHE)
-    foreach(cache) do (k, v)
-        tz = first(v)
-        cache = tz isa FixedTimeZone ? TimeZones._FTZ_CACHE : TimeZones._VTZ_CACHE
-        setindex!(cache, v, k)
-    end
+function with_tz_cache(f, cache::TimeZones.TimeZoneCache)
+    old_cache = deepcopy(TimeZones._TZ_CACHE)
+    copy!(TimeZones._TZ_CACHE, cache)
 
     try
         return f()
     finally
-        copy!(TimeZones._FTZ_CACHE, old_ftz_cache)
-        copy!(TimeZones._VTZ_CACHE, old_vtz_cache)
+        copy!(TimeZones._TZ_CACHE, old_cache)
     end
 end