--- a/test/types/timezone.jl
+++ b/test/types/timezone.jl
@@ -1,15 +1,5 @@
 using TimeZones: Class
 
-@testset "TimeZone allocations" begin
-    tz = TimeZone("UTC")  # run once for compilation and to populate cache
-    @assert tz isa FixedTimeZone
-    @test @allocations(TimeZone("UTC")) == 0
-
-    tz = TimeZone("America/Winnipeg")  # populate cache
-    @assert tz isa VariableTimeZone
-    @test @allocations(TimeZone("America/Winnipeg")) == 2
-end
-
 @testset "istimezone" begin
     @test istimezone("Europe/Warsaw")
     @test istimezone("UTC+02")
@@ -34,4 +24,16 @@ end
     @test TimeZone("Etc/GMT", Class(:LEGACY)) == FixedTimeZone("Etc/GMT", 0)
     @test TimeZone("Etc/GMT+12", Class(:LEGACY)) == FixedTimeZone("Etc/GMT+12", -12 * 3600)
     @test TimeZone("Etc/GMT-14", Class(:LEGACY)) == FixedTimeZone("Etc/GMT-14", 14 * 3600)
+end
+
+@testset "allocations" begin
+    tz = TimeZone("UTC")  # Trigger compilation and ensure the cache is populated
+    @test tz isa FixedTimeZone
+    @test @allocations(TimeZone("UTC")) == 0
+    @test @allocations(istimezone("UTC")) == 0
+
+    tz = TimeZone("America/Winnipeg")  # Trigger compilation and ensure the cache is populated
+    @test tz isa VariableTimeZone
+    @test @allocations(TimeZone("America/Winnipeg")) == 2
+    @test @allocations(istimezone("America/Winnipeg")) == 1
 end