Closes https://github.com/JuliaTime/TimeZones.jl/issues/422

This is a rebase/reimplementation of #423  onto `master` .
(FYI @nickrobinson251, I'm doing this after seeing [your comment](https://github.com/JuliaTime/TimeZones.jl/pull/423#issuecomment-1726079158)!)

I started by trying to rebase your PR properly, but in the end it was easier to cherry-pick your commits that still applied (just the tests), then re-implement the rest. I'm sorry that this results in this new PR, but wasn't sure how else to proceed.

One note -- I find that it now takes two allocations to retrieve a `VariableTimeZone` from the cache, rather than 1. I spent some time with the alllocation tracker trying to understand this with no luck, so I've updated the test expectation accordingly. This isn't any worse than the current behaviour on `master`, and in any case the main thing is that retrieving a `FixedTimeZone` is now allocation free :)