@omus 2024-05-23
Baseline timings for my machine. I'll need to resolve the conflict yet.

Julia 1.10.0

Current `master` (ac06a91)

```julia
julia> @btime TimeZone("America/Winnipeg");
  41.246 ns (2 allocations: 64 bytes)

julia> @btime TimeZone("UTC");
  40.868 ns (2 allocations: 64 bytes)
```

This PR (2b0b76e)

```julia
julia> @btime TimeZone("UTC");
  17.994 ns (0 allocations: 0 bytes)

julia> @btime TimeZone("America/Winnipeg");
  28.475 ns (2 allocations: 96 bytes)
```