@omus 2024-05-23
These changes are also a step in the right direction for reducing the initial cache loading time

This PR 67f2c3f
```juliq
julia> @btime TimeZones._reload_tz_cache(TimeZones._COMPILED_DIR[]);
  20.795 ms (313574 allocations: 12.52 MiB)
```

Current `master` ac06a91
```julia
julia> @btime TimeZones._reload_tz_cache(TimeZones._COMPILED_DIR[]);
  22.808 ms (313621 allocations: 12.52 MiB)
```